+++ src/language-graphql/printer-graphql.js
@@ -14,8 +14,11 @@ import getVisitorKeys from "./get-visitor-keys.js";
 import { locStart } from "./loc.js";
 import { massageAstNode } from "./massage-ast/index.js";
 import { insertPragma } from "./pragma.js";
+import { printArguments } from "./print/arguments.js";
 import { printDescription } from "./print/description.js";
+import { printDirectives } from "./print/directives.js";
 import { printSequence } from "./print/sequence.js";
+import { printVariableDefinitions } from "./print/variable-definitions.js";
 
 function genericPrint(path, options, print) {
   const { node } = path;
@@ -425,25 +428,6 @@ function genericPrint(path, options, print) {
   }
 }
 
-function printDirectives(path, print) {
-  const { node } = path;
-
-  if (!isNonEmptyArray(node.directives)) {
-    return "";
-  }
-
-  const printed = join(line, path.map(print, "directives"));
-
-  if (
-    node.kind === "FragmentDefinition" ||
-    node.kind === "OperationDefinition"
-  ) {
-    return group([line, printed]);
-  }
-
-  return [" ", group(indent([softline, printed]))];
-}
-
 function canAttachComment(node /* , ancestors */) {
   return node.kind !== "Comment";
 }
@@ -455,25 +439,6 @@ function printComment({ node: comment }) {
 
   /* c8 ignore next */
   throw new Error("Not a comment: " + JSON.stringify(comment));
-}
-
-function printVariableDefinitions(path, print) {
-  const { node } = path;
-  if (!isNonEmptyArray(node.variableDefinitions)) {
-    return "";
-  }
-  return group([
-    "(",
-    indent([
-      softline,
-      join(
-        [ifBreak("", ", "), softline],
-        path.map(print, "variableDefinitions"),
-      ),
-    ]),
-    softline,
-    ")",
-  ]);
 }
 
 function hasPrettierIgnore(path) {
